do the same thing with context api
