+++ src/Components/TaskDetails.js
@@ -1,42 +1,38 @@
-import { useEffect, useState } from "react";
-import { useParams, useNavigate } from "react-router-dom";
-import { Link } from "react-router-dom";
+import { useEffect } from "react";
+import { useParams, Link } from "react-router-dom";
+import { useTasks } from "./TaskContext";
 import { TASK_API_URL_SLASH } from "../constant";
 
 const TaskDetails = () => {
-    const { taskid } = useParams();
-    const [taskdata, setTaskData] = useState(null);
-    
+  const { taskid } = useParams();
+  const { currentTask, setCurrentTask } = useTasks();
 
-    useEffect(() => {
-        fetch(TASK_API_URL_SLASH + taskid)
-            .then((res) => {
-                return res.json();
-            })
-            .then((resp) => {
-                setTaskData(resp);
-            })
-            .catch((err) => {
-                console.log(err.message);
-            });
-    }, []);
+  useEffect(() => {
+    setCurrentTask(null);  // reset before fetching new task
+    fetch(TASK_API_URL_SLASH + taskid)
+      .then(res => res.json())
+      .then(data => setCurrentTask(data))
+      .catch(err => console.log(err.message));
+  }, [taskid, setCurrentTask]);
 
-    return (
-        <div className="container">
-            {taskdata &&
-                <div className="card">
-                    <div className="card-header">
-                        <h2>Task Details</h2>
-                    </div>
-                    <div className="card-body" style={{textAlign:"left"}}>
-                        <h3>Task Name: {taskdata.name} ({taskdata.id})</h3>
-                        <p>Description: {taskdata.description}</p>
-                        <Link className = "btn btn-remove" to ="/">Back to task list</Link>
-                    </div>
-                </div>
-              }
+  if (!currentTask) {
+    return <div>Loading task details...</div>;
+  }
+
+  return (
+    <div className="container">
+      <div className="card">
+        <div className="card-header">
+          <h2>Task Details</h2>
         </div>
-    );
+        <div className="card-body" style={{ textAlign: "left" }}>
+          <h3>Task Name: {currentTask.name} ({currentTask.id})</h3>
+          <p>Description: {currentTask.description}</p>
+          <Link className="btn btn-remove" to="/">Back to task list</Link>
+        </div>
+      </div>
+    </div>
+  );
 };
 
 export default TaskDetails;
