+++ src/Components/TaskList.js
@@ -1,80 +1,26 @@
-import { useEffect, useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { confirmAlert } from "react-confirm-alert";
-import 'react-confirm-alert/src/react-confirm-alert.css';
-import { TASK_API_URL_SLASH } from "../constant";
+import { useContext, useEffect } from "react";
+import { Link } from "react-router-dom";
 import { useSwipeable } from "react-swipeable";
+import { TASK_API_URL_SLASH } from "../constant";
+import { useTasks } from './TaskContext';
+
+
 
 const TaskList = () => {
-    const [taskdata, setTaskData] = useState(null);
-    const [loadingTaskId, setLoadingTaskId] = useState(null);
-    const [crossOverTask, setCrossOverTask] = useState([]);
+    const { taskdata, setTaskData, loadingTaskId, setLoadingTaskId, crossOverTask, setCrossOverTask, LoadDetail, LoadEdit, RemoveFunction, handleDeleteDirect, tapTaskCross } = useTasks();
 
-    const navigate = useNavigate();
-
-    const LoadDetail = (id) => {
-        navigate('/task/details/' + id);
-    };
-
-    const LoadEdit = (id) => {
-        navigate('/task/edit/' + id);
-    };
-
-    const RemoveFunction = (id) => {
-        setLoadingTaskId(id);
-        confirmAlert({
-            title: 'Confirm to delete',
-            message: 'Are you sure you want to delete this task?',
-            buttons: [
-                {
-                    label: 'Yes',
-                    onClick: () => {
-                        fetch(TASK_API_URL_SLASH + id, { method: "DELETE" })
-                            .then((res) => {
-                                setTaskData((prev) => prev.filter(task => task.id !== id));
-                                alert('Deleted Successfully');
-                                setLoadingTaskId(null);
-                            })
-                            .catch((err) => {
-                                console.log(err.message);
-                                setLoadingTaskId(null);
-                            });
-                    }
-                },
-                {
-                    label: 'No',
-                    onClick: () => setLoadingTaskId(null)
-                }
-            ]
-        });
-    };
-
-    const handleDeleteDirect = (id) => {
-        setLoadingTaskId(id);
-        fetch(TASK_API_URL_SLASH + id, { method: "DELETE" })
-            .then(() => {
-                setTaskData((prev) => prev.filter(task => task.id !== id));
-                setLoadingTaskId(null);
-            })
-            .catch((err) => {
-                console.log(err.message);
-                setLoadingTaskId(null);
-            });
-    };
-
-    const tapTaskCross = (id) => {
-        if (crossOverTask.includes(id)) {
-            setCrossOverTask(prev => prev.filter(taskId => taskId !== id));
-        } else {
-            setCrossOverTask(prev => [...prev, id]);
-        }
-    };
+    useEffect(() => {
+        fetch(TASK_API_URL_SLASH)
+            .then((res) => res.json())
+            .then((resp) => setTaskData(resp))
+            .catch((err) => console.log(err.message));
+    }, []);
 
     const TaskRow = ({ item }) => {
         const handlers = useSwipeable({
-            onSwipedLeft: () => handleDeleteDirect(item.id),
-            preventDefaultTouchmoveEvent: true,
-            trackMouse: true
+            onSwipedLeft: () => tapTaskCross(item.id),
+            onSwipedRight: () => tapTaskCross(item.id),
+            preventScrollOnSwipe: true,
         });
 
         const isTapped = crossOverTask.includes(item.id);
@@ -95,13 +41,6 @@ const TaskList = () => {
         );
     };
 
-    useEffect(() => {
-        fetch(TASK_API_URL_SLASH)
-            .then((res) => res.json())
-            .then((resp) => setTaskData(resp))
-            .catch((err) => console.log(err.message));
-    }, []);
-
     return (
         <div className="container">
             <div className="card">
@@ -110,6 +49,7 @@ const TaskList = () => {
                 </div>
                 <div className="card-body">
                     <div><Link to="task/create" className="btn btn-green">Add New Task</Link></div>
+
                     <table className="my-table">
                         <thead>
                             <tr>
@@ -128,6 +68,7 @@ const TaskList = () => {
                             ))}
                         </tbody>
                     </table>
+
                 </div>
             </div>
         </div>
